Set up new Gradle project structure with Kotlin and LibGDX
Added new Gradle tasks for multi-module Kotlin project setup with LibGDX. In detail:
- Created build scripts for root, app, core and gdx modules.
- Configured Kotlin JVM version and added necessary plugins.
- Set up .idea project settings and code style configurations.
- Adjusted .gitignore for Gradle and IntelliJ-based projects.
- Implemented basic Gradle wrapper batch and shell scripts.
- Included sample Kotlin classes to verify project setup.
Also set-up a project structure consisting out of 3 modules:
1. app
2. core
3. gdx
Where gdx is a wrapper module to decouple the actual application from libgdx, core is the module containing the game logic and systems, and app as the module responsible for execution.

This commit sets the foundation for further development by defining the basic project structure and dependencies.