:recycle: Refactor game rendering logic, introduce GameAdapter
Removed the Game class to provide better structure and ease of adaptability. Now game rendering logic is handled through an abstract class, GameAdapter, with a forced implementation of a 'renderScene' method. This allows for cleaner game-specific implementations. Furthermore, a new game, Stranded, was created to demonstrate the new structure and implementation mechanism.