:sparkles: Add clearScreen util and integrate into Stranded class
Two new Kotlin files have been added: RenderingUtils.kt and Color.kt. The former includes a function clearScreen to clear the game screen with a customizable color, defaulting to black as specified in the new Color data class in Color.kt. Also, clearScreen() has been integrated into the renderScene method of the Stranded class to clear the screen before a scene is rendered. This is meant to prevent artifacts from previous frames persisting on the display.