:recycle: Refactor game rendering to improve game performance
In the previous logic, spriteBatch was replicated in every game instance. This could potentially lead to memory inefficiency and degrade game performance. To address this, the rendering logic has been moved to a base class GameAdapter which manages the life cycle of SpriteBatch. In addition, functions to interact with sprites have been abstracted to provide more flexibility in dealing with different game mechanics, hence improving the modularity of the game. Also, the overall code readability improved, making future developments easier.