:sparkles: Add Graphics wrapper to centralize viewport dimensions
The Graphics wrapper has been introduced to centralize the viewport dimensions throughout the game. The newly created class, Graphics, has width and height as properties and it is passed during the initialization of the GameAdapter, which is reflected in any class extending from this GameAdapter. This ensures that viewport dimensions are not hard-coded, improving maintainability and scalability.