:bricks: Update Gradle settings and introduce Detekt plugin
This commit introduces the GradleMigrationSettings component in the gradle.xml file with migrationVersion set at "1". The objective was to ensure smooth Gradle project migration.

In addition, a new detekt.xml file has been added to integrate the Detekt static code analysis plugin for improving code quality. The "enableForProjectResult" option is however currently set to "Declined".

A minor change in the misc.xml file removes the xml version tag, a superfluous declaration, for a cleaner code base.